v4 - Added an introduction and Improvement of report outlook
Included an introduction in the report, cleaner code and more readable markdown cells. The purpose of the commit is purely for the end report to look more professional, readable and since the end of the challenge is near on Kaggle.

To further add. In v3 commit, I had not mentioned that the SVC classification model had also received hyperparameter tuning...The overall changes mentioned in the v3 commit (description) resulted in an increase of the F1_score, Precision and Accuracy for all the models.