Added in Model analysis & Improved report outlook
Added in Model analysis of all models performances. The Analysis is done to justify my reasononing for selecting the Linear SVC model as my chosen model to be used to predict and classify the full Test data set.

I also took the liberty to further improve the report oulook with cleaner code, added markdowns for more and better explanations and reasonings.