v3: Improvement of report outlook and hyperparameter tuning.
Improvement of report: Improved the aspect of the report, better/cleaner code and included a code to train all imported models (in one loop) to obtain/choose the best f1_score.

Hyperparameter tuning: In regards to the imported classification models, I tuned some of the parameters to improve end result, added in C=2.1, loss = 'hinge' for LinearSVC. Added in tuning for Random forest, and Decision Tree models as a first for this version commit.