V5 - Added in Model analysis & Improved report outlook
Added in model analysis based on all model performance. The analysis is done to jusify my reasoning for selecting the LinearSVC model as my chosen model to be used to predict and classify the full Test data set.

I also took the liberty to further improve the outlook of the report with cleaner code, added markdowns for more and better explanations.